Updated the Readme file to accept template id
Added the parameter to accept template id in settings.
Goto Advance Usage Example to check the usage of template id.

One can upload the template in Pepipost and get the template id. This template id can be passed here in the SDK. This helps in reducing the data transfer over network.